renamed login to auth and added confirm-email function
