+++ lib/auth/register.dart
