+++ lib/main.dart
@@ -2,8 +2,8 @@ import "package:flutter/material.dart";
 import "package:flutter/services.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:mathlympics/leaderboard.dart";
-import "package:mathlympics/login/page.dart";
-import "package:mathlympics/login/register.dart";
+import "package:mathlympics/auth/login.dart";
+import "package:mathlympics/auth/register.dart";
 import "package:mathlympics/normal_game_screen.dart";
 import "package:mathlympics/play_screen.dart";
 import "package:supabase_flutter/supabase_flutter.dart";
